Add Google DLP validation to user prompts
This commit adds some validation to the user query to ensure they do not
submit Personal Identifying Information. It utilises the google-cloud-dlp gem
to identify possible PII and stops the query from being sent to the govuk chat api.
Currently it is configured to identify names and US state, but we can add more infoTypes
for it to search for. This is currently a proof-of-concept as part of a spike
into detecting PII.

The infoTypes we can configure it to look for are described here
https://cloud.google.com/dlp/docs/infotypes-reference