Do not allow magic links to be used more than once
Prevents magic links from being used more than once.
This aims to deter the prototype link from being shared widely.
This is because the product will be used for a user research pilot
and we do not intend for it to go live. We do not want it to be
used by people who are not aware of this. This will help
to reduce the risk but will not eliminate it entirely.

The link expiry and session time out default values are also
defined here. Initially they have been set to two days.