Bug fix
There are now potentially several forms on the page and they all contain a hidden js_enabled input field that is sent with the request to the server. Accordingly, all these hidden js_enabled input fields need to be switched to value: true, hence the refactor of setJSEnabled(). It needs to be called in several places so that it handles all mutations of the DOM (i.e. avoids a situation where JS is available but doesn't set js_enabled to true)