Fix PII checks not working after submitting feedback
This was caused by two things. Firstly, the PII function was hooking into the first submit button it could find. The first submit button on the page is now the Feedback forms under each message. Therefore it was checking for PII when a feedback button was pressed, not the chat 'Send' button. Secondly, this was because it would only add the PII detection when a new message triggered a turbo reload. When you submit feedback, no new messages are added to the DOM, so the PII detection was not being reintroduced. Now, the function specifically looks for the chat's 'send' button, and will add the PII detector after every turbo refresh of the DOM.