Fix issues with new chats and the loading icon
Looks like display inline/inline-block/block inconsistencies was causing issues. This fixes the first chat message not being wide enough and then suddenly becomes wide enough once a long chat response appears. The second issue is that the loading animation is within govuk-chat-icon, and this was set to inline, which broke the loading animation.