Use a configurable prefix for chat id
This change has been made so that we can have two instances of this
application running with different prefixes.

The use case for this is: we have a version of the prototype running
for HMRC evaluation with "proto" as a prefix, and a different one for
demoing to stakeholders with one of "demo".